requirements.txt
yfinance
requests
beautifulsoup4
python-dotenv
anthropic
pandas
numpy
lxml
ta